Fix tests
The tests have been acting up (specifically pylint, but due to a bug I
introduced by not properly importing things the python test as well.)

This commit addresses those things.

One note is that I ignored the `len-as-condition` pylint error. Because
I don't see a reason to refactor key parts of the app for what is a
pretty minor error.