Use _test_login_success to simplify test_login.py.
There was a lot of repeated code in `test_login.py`, and I was going to have to add even more in a test of #351. So, this commit pulls out the common code for succesful logins into `_test_login_success`, which is called by other tests. `test_custom_session_success` goes away entirely because `test_default_domain_success` is almost identical.