Strip leading v from api_version (#522)
* Strip leading v from api_version

Strip leading v from version before templating into url. Including the 
leading v in the api_version in url results in the server responding with 
API V7.0 in the server_url

* fix string formatting

Fix string formatting

* Replace silent version correction with Exception

Raise an exception when sf_version is specified with a leading 'v', since 
specifying a version with the leading 'v' leads to surprising behavior in
Salesforce.