Updated __getattr__ function to create SFType object with object_pairs_hook (#634)
* Bug/pyjwt dependency (#610)

* Update setup.py

Update pyjwt dependency

* Update setup.py

include cryptography for requirement for pyjwt

* Update setup.py

include pyjwt[crypto] in install requirements

* Update setup.py

remove [crypto] from pyjwt[crypto] in install_requires as causes travis-ci failure to build

* Updated __getattr__ function to create SFType object with object_pairs_hook as input

Right now let us say that the user creates a SFType object using __getattr__ function which is present in the class SalesForce. If the SalesForce object is initiated using object_pairs_hook being dict or print. In that case the user would want the response from SFType to be same as that object as well. So I have added object_pairs_hook in the SFType object which is being called in the __getattr__ function of the SalesForce Class

* Updated __getattr__ function to create SFType object with object_pairs_hook

Right now let us say that the user creates a SFType object using getattr function which is present in the class SalesForce. If the SalesForce object is initiated using object_pairs_hook being dict or print. In that case the user would want the response from SFType to be same as that object as well. So I have added object_pairs_hook in the SFType object which is being called in the getattr function of the SalesForce Class

* removed whitespaces

* pylint

---------

Co-authored-by: JonWobken <49450080+jon-wobken@users.noreply.github.com>